move getNamespace
