+++ libs/evm-protocols/src/index.ts
@@ -1,5 +1,6 @@
 export * from './abis';
 export * as commonProtocol from './common-protocol';
+export * from './common-protocol/contractHelpers/namespace';
 export * from './common-protocol/contractHelpers/stake';
 export * from './event-registry/eventRegistry';
 export * from './event-registry/eventSignatures';
