+++ libs/model/src/services/commonProtocol/contractHelpers.ts
@@ -7,50 +7,6 @@ import { AbiItem } from 'web3-utils';
 import { Balances, TokenAttributes, getBalances } from '../tokenBalanceCache';
 
 /**
- * Retrieves a namespace.
- * @param rpcNodeUrl Note this MUST be a private_url with no associated whitelist.
- * @param namespace
- * @param factoryAddress
- */
-export const getNamespace = async (
-  rpcNodeUrl: string,
-  namespace: string,
-  factoryAddress: string,
-): Promise<string> => {
-  const web3 = new Web3(rpcNodeUrl);
-  const factory = new web3.eth.Contract(
-    [
-      {
-        inputs: [
-          {
-            internalType: 'bytes32',
-            name: '',
-            type: 'bytes32',
-          },
-        ],
-        stateMutability: 'view',
-        type: 'function',
-        name: 'getNamespace',
-        outputs: [
-          {
-            internalType: 'address',
-            name: '',
-            type: 'address',
-          },
-        ],
-      },
-    ] as AbiItem[],
-    factoryAddress,
-  );
-
-  const hexString = web3.utils.utf8ToHex(namespace);
-  const activeNamespace = await factory.methods
-    .getNamespace(hexString.padEnd(66, '0'))
-    .call();
-  return String(activeNamespace);
-};
-
-/**
  * gets the balances of an id for multiple addresses on a namespace
  * @param namespaceAddress the contract address of the deployed namespace
  * @param tokenId ERC1155 id(ie 0 for admin token, default 2 for CommunityStake)
